+++ server/src/test/service/LocaisService.test.js
@@ -12,6 +12,32 @@ import mongoose from 'mongoose';
  * @author Estácio Pereira.
  */
 describe('LocaisServiceTest', () => {
+        function getMockImagem(tipo, tamInvalido = false) {
+            let conteudo = `data:${tipo};base64,asdasdasdasdasdasdasdasddassasdasda`;
+            if (tamInvalido) {
+                return {conteudo: gerarConteudoMaiorLimite(conteudo)};
+            }
+            return {conteudo};
+        }
+
+        /**
+         * Cria um Mock de String para o conteúdo.
+         * O mock vai retornar no método substrig um início de string válido
+         * para ser utilizado no validador.
+         * O método split vai retornar apenas um objeto com a propriedade length
+         * com tamanho maior que o máximo (16 MB).
+         * 
+         * @param {String} conteudo conteudo atual da imagem;
+         */
+        function gerarConteudoMaiorLimite(conteudo) {
+            let tamMaiorLimite = 16*1000*1005;
+            return { 
+                substring: (a, b) => conteudo.substring(a, b), 
+                split: () => { 
+                    return {  length: tamMaiorLimite }; 
+                } 
+            };
+        }
 
     describe('cadastrarLocal deve', () => {
         it('não salvar um local nulo', () => {
@@ -42,6 +68,81 @@ describe('LocaisServiceTest', () => {
                     expect(local.capacidade).to.be.eql(mockLocal.capacidade);
                     expect(local.funcionamento).to.be.eql(mockLocal.funcionamento);
                 });
+        });
+
+        it('não salvar um local com imagem de tipo inválido', () => {
+            const mockLocal = LocaisMock.getLocal();
+            const imagemInvalida = getMockImagem("file/pdf");
+
+            mockLocal.imagens.push(imagemInvalida);
+            return LocaisService.cadastrarLocal(mockLocal)
+                .then(local => assert.fail('Não deveria salvar imagem com tipo inválido.'))
+                .catch(err => {
+                    expect(err).to.exist;
+                    expect(err).to.be.equal("Tipo de imagem não suportado, deve ser jpg, jpeg, png, bitmap ou gif.");
+                });
+        });
+
+        it('não salvar um local com imagem de tamanho maior que 16MB e tipo inválido', () => {
+            const mockLocal = LocaisMock.getLocal();
+            const imagemInvalida = getMockImagem("file/pdf", true);
+
+            mockLocal.imagens.push(imagemInvalida);
+            return LocaisService.cadastrarLocal(mockLocal)
+                .then(local => assert.fail('Não deveria salvar imagem com tipo e tamanho inválido.'))
+                .catch(err => {
+                    expect(err).to.exist;
+                    expect(err).to.be.equal("O tamanho máximo suportado é 16MB. Tipo de imagem não suportado, deve ser jpg, jpeg, png, bitmap ou gif.");
+                });
+        });
+
+        it('não salvar um local com imagem de tamanho maior que 16MB', () => {
+            const mockLocal = LocaisMock.getLocal();
+            const imagemInvalida = getMockImagem("image/png", true);
+            
+            mockLocal.imagens.push(imagemInvalida);
+            return LocaisService.cadastrarLocal(mockLocal)
+                .then(local => assert.fail('Não deveria salvar imagem com tamanho inválido.'))
+                .catch(err => {
+                    expect(err).to.exist;
+                    expect(err).to.be.equal("O tamanho máximo suportado é 16MB. ");
+                });
+        });
+
+        it(' salvar um local com imagem de tipo válido jpeg', () => {
+            const mockLocal = LocaisMock.getLocal();
+            const imagemValida = getMockImagem("image/jpeg");
+
+            mockLocal.imagens.push(imagemValida);
+            return LocaisService.cadastrarLocal(mockLocal)
+                .then(local => {
+                    expect(local.imagens).to.have.length(1);
+                })
+                .catch(err => assert.fail('Deveria salvar imagem com tipo válido.'));
+        });
+
+        it(' salvar um local com imagem de tipo válido bitmap', () => {
+            const mockLocal = LocaisMock.getLocal();
+            const imagemValida = getMockImagem("image/bitmap");
+            
+            mockLocal.imagens.push(imagemValida);
+            return LocaisService.cadastrarLocal(mockLocal)
+                .then(local => {
+                    expect(local.imagens).to.have.length(1);
+                })
+                .catch(err => assert.fail('Deveria salvar imagem com tipo válido.'));
+        });
+
+        it(' salvar um local com imagem de tipo válido png', () => {
+            const mockLocal = LocaisMock.getLocal();
+            const imagemValida = getMockImagem("image/png");
+
+            mockLocal.imagens.push(imagemValida);
+            return LocaisService.cadastrarLocal(mockLocal)
+                .then(local => {
+                    expect(local.imagens).to.have.length(1);
+                })
+                .catch(err => assert.fail('Deveria salvar imagem com tipo válido.'));
         });
     });
 
@@ -124,6 +225,42 @@ describe('LocaisServiceTest', () => {
                     expect(localAtualizado._id).to.be.eql(localPersistido._id);
                     expect(localAtualizado).to.be.eql(novoLocal);
                 });
+        });
+
+        it('não atualizar o local caso exista imagem maior que 16mb', () => {
+            const novoLocal = _.clone(localPersistido);
+            novoLocal.imagens.push(getMockImagem('image/jpg', true));
+
+            return LocaisService.atualizarLocal(localPersistido._id, novoLocal)
+                .then(localAtualizado => assert.fail("Não deveria ser possível atualizar local com imagem maior que 16mb."))
+                .catch(err => {
+                    expect(err).to.exist;
+                    expect(err).to.be.equal("O tamanho máximo suportado é 16MB. ");
+            });
+        });
+
+        it('não atualizar o local caso exista imagem maior que 16mb e tipo inválido', () => {
+            const novoLocal = _.clone(localPersistido);
+            novoLocal.imagens.push(getMockImagem('file/pdf', true));
+
+            return LocaisService.atualizarLocal(localPersistido._id, novoLocal)
+                .then(localAtualizado => assert.fail("Não deveria ser possível atualizar local com imagem maior que 16mb."))
+                .catch(err => {
+                    expect(err).to.exist;
+                    expect(err).to.be.equal("O tamanho máximo suportado é 16MB. Tipo de imagem não suportado, deve ser jpg, jpeg, png, bitmap ou gif.");
+            });
+        });
+
+        it('não atualizar o local caso exista imagem com tipo inválido', () => {
+            const novoLocal = _.clone(localPersistido);
+            novoLocal.imagens.push(getMockImagem('file/pdf'));
+
+            return LocaisService.atualizarLocal(localPersistido._id, novoLocal)
+                .then(localAtualizado => assert.fail("Não deveria ser possível atualizar local com imagem maior que 16mb."))
+                .catch(err => {
+                    expect(err).to.exist;
+                    expect(err).to.be.equal("Tipo de imagem não suportado, deve ser jpg, jpeg, png, bitmap ou gif.");
+            });
         });
     });